Separate utilities into module
